Moved t_placer_costs and t_annealing_state and related routines to placer_util.* files.
